+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.29:
 
+* Remove support for recently added "embedded.glassfish.specs" system setting in as the implementation is unlikely
+  to be usable.
 * Add utility method GlassFishContainer.addSpecToClasspath() that can add maven style dependencies to classpath when
   creating GlassFish.
 * Add guard against adding file that does not exist to GlassFishContainer's classpath.